Updating interfaces Fixing dialog up checks formatting the hooks Fixing doAction interface Fixing login handler fixing login response ADding the ability to minimize on start up with the "-m" flag Fixing analizer

+++ src/com/p3achb0t/client/injection/InstanceManager.kt
@@ -395,7 +395,7 @@ class InstanceManager(val client: Any): Logging() {
     }
 
     //.doActionCallback(int, int, int, int, java.lang.String, java.lang.String, int, int, int)'
-    fun doActionCallback(argument0: Int, argument1: Int, argument2: Int, argument3: Int, action: String, targetName: String, mouseX: Int, mouseY: Int, argument8: Byte) {
+    fun doActionCallback(argument0: Int, argument1: Int, argument2: Int, argument3: Int, action: String, targetName: String, mouseX: Int, mouseY: Int, argument8: Int) {
         logger.info("argument0:$argument0, argument1:$argument1, argument2:$argument2, argument3:$argument3, action:$action, targetName:$targetName, mouseX:$mouseX, mouseY:$mouseY, argument8:$argument8")
     }
